requirements.txt
pypdf>=3.0.0
python-pptx>=0.6.21
pywin32>=305; sys_platform == 'win32'
